# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PaySim Sandbox API",
    "description": "\nIsolated sandbox that simulates the full payment lifecycle. **No real money, no bank connections.**\n\n**Lifecycle:** `pending \u2192 processing \u2192 success | failed | timeout`, then `success \u2192 refunded`.\n\n**Security:** Bearer API key on every request, `Idempotency-Key` on every POST that changes money state,\nstrict schema validation, and a rate limit of 60 requests/minute per API key.\n\n**Errors** always use `{\"error\": {\"code\", \"message\", \"details\"}}`:\n\n| HTTP | code | When |\n|---|---|---|\n| 400 | `IDEMPOTENCY_KEY_MISSING` | POST without an `Idempotency-Key` header |\n| 401 | `UNAUTHORIZED` | Missing or invalid Bearer API key |\n| 404 | `PAYMENT_NOT_FOUND` | Unknown payment id (or another merchant's) |\n| 409 | `IDEMPOTENCY_KEY_CONFLICT` | Same `Idempotency-Key` reused with a different body |\n| 409 | `INVALID_STATE_TRANSITION` | e.g. refunding a payment that is not `success` |\n| 422 | `VALIDATION_ERROR` | Invalid amount, currency, reference, simulation_mode or extra fields |\n| 429 | `RATE_LIMITED` | More than 60 requests/minute (see `Retry-After`) |\n| 502 | `AI_ERROR` | Gemini request failed or is overloaded |\n| 503 | `AI_UNAVAILABLE` | `GEMINI_API_KEY` is not configured |\n",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/payments": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create a sandbox payment",
        "operationId": "create_payment_api_v1_payments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Unique key per operation (e.g. a UUID). Repeating it returns the original response.",
              "title": "Idempotency-Key"
            },
            "description": "Unique key per operation (e.g. a UUID). Repeating it returns the original response."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing Idempotency-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency-Key reused with a different body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid amount, currency, reference or simulation_mode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/{transaction_id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Get payment status",
        "operationId": "get_payment_api_v1_payments__transaction_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Payment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/{transaction_id}/refund": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Fully refund a successful payment",
        "operationId": "refund_payment_api_v1_payments__transaction_id__refund_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Unique key per operation (e.g. a UUID). Repeating it returns the original response.",
              "title": "Idempotency-Key"
            },
            "description": "Unique key per operation (e.g. a UUID). Repeating it returns the original response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing Idempotency-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Payment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Payment is not in 'success' status, or Idempotency-Key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sandbox/smoke-test": {
      "post": {
        "tags": [
          "Sandbox"
        ],
        "summary": "Run the full payment lifecycle smoke test",
        "operationId": "smoke_test_api_v1_sandbox_smoke_test_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmokeTestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ai/debug": {
      "post": {
        "tags": [
          "AI Assistant"
        ],
        "summary": "Explain a failed API call",
        "operationId": "debug_api_v1_ai_debug_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebugRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebugSuggestion"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Gemini request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Gemini is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ai/audit-summary/{transaction_id}": {
      "post": {
        "tags": [
          "AI Assistant"
        ],
        "summary": "Plain-English audit summary of a transaction",
        "operationId": "audit_summary_api_v1_ai_audit_summary__transaction_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditSummaryResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "description": "Gemini request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Gemini is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuditSummaryResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "summary"
        ],
        "title": "AuditSummaryResponse"
      },
      "CreatePaymentRequest": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Amount",
            "description": "Positive amount with at most 2 decimal places.",
            "examples": [
              "100.00"
            ]
          },
          "currency": {
            "type": "string",
            "enum": [
              "LYD",
              "USD",
              "EUR"
            ],
            "title": "Currency",
            "description": "ISO 4217 code (uppercase).",
            "examples": [
              "LYD"
            ]
          },
          "reference": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_\\-]+$",
            "title": "Reference",
            "description": "Merchant order reference.",
            "examples": [
              "ORDER-1001"
            ]
          },
          "simulation_mode": {
            "type": "string",
            "enum": [
              "force_success",
              "force_failure",
              "force_timeout"
            ],
            "title": "Simulation Mode",
            "description": "Controls the simulated outcome.",
            "default": "force_success"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount",
          "currency",
          "reference"
        ],
        "title": "CreatePaymentRequest"
      },
      "DebugRequest": {
        "properties": {
          "method": {
            "type": "string",
            "maxLength": 10,
            "title": "Method",
            "examples": [
              "POST"
            ]
          },
          "path": {
            "type": "string",
            "maxLength": 300,
            "title": "Path",
            "examples": [
              "/api/v1/payments"
            ]
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          },
          "body": {
            "title": "Body"
          },
          "response_status": {
            "type": "integer",
            "title": "Response Status"
          },
          "response_body": {
            "title": "Response Body"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "method",
          "path",
          "response_status"
        ],
        "title": "DebugRequest",
        "description": "A failed API call to be analysed by Gemini."
      },
      "DebugSuggestion": {
        "properties": {
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "cause": {
            "type": "string",
            "title": "Cause"
          },
          "fix": {
            "type": "string",
            "title": "Fix"
          },
          "corrected_payload": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Payload",
            "description": "Corrected JSON body as a string, if relevant."
          }
        },
        "type": "object",
        "required": [
          "problem",
          "cause",
          "fix"
        ],
        "title": "DebugSuggestion"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PaymentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "object": {
            "type": "string",
            "const": "payment",
            "title": "Object",
            "default": "payment"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "simulation_mode": {
            "type": "string",
            "title": "Simulation Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "success",
              "failed",
              "timeout",
              "refunded"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "refunded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refunded At"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/StatusChange"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "id",
          "amount",
          "currency",
          "reference",
          "simulation_mode",
          "status",
          "created_at",
          "updated_at",
          "refunded_at",
          "history"
        ],
        "title": "PaymentResponse"
      },
      "SmokeStep": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "passed",
              "failed",
              "skipped"
            ],
            "title": "Status"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "name",
          "status",
          "duration_ms",
          "detail"
        ],
        "title": "SmokeStep"
      },
      "SmokeTestResponse": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          },
          "total_ms": {
            "type": "integer",
            "title": "Total Ms"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/SmokeStep"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "transaction_id",
          "total_ms",
          "steps"
        ],
        "title": "SmokeTestResponse"
      },
      "StatusChange": {
        "properties": {
          "from_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Status"
          },
          "to_status": {
            "type": "string",
            "title": "To Status"
          },
          "event": {
            "type": "string",
            "title": "Event"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "from_status",
          "to_status",
          "event",
          "at"
        ],
        "title": "StatusChange"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Merchant API key, e.g. sk_test_...",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "https://127.0.0.1:8000",
      "description": "Local sandbox (HTTPS)"
    }
  ]
}